+++ app/src/main/res/drawable/bg_image_friend.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layer-list xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:drawable="@drawable/img_summer01" />
-    <item android:drawable="@drawable/img_friend_circle" />
+    <item android:drawable="@drawable/img_summer" />
+    <item android:drawable="@drawable/bg_friend_circle" />
 </layer-list>
